requirements.txt
PyYAML==5.3
easydict==1.9
numpy==1.18.1
pandas==1.0.1
matplotlib==3.1.3
colorlog=4.1.0